feat: partial implementation of production order logic

+++ src/main/java/com/gestone/gestone_api/domain/order/MarbleshopOrderDTO.java
@@ -4,13 +4,14 @@ import com.gestone.gestone_api.domain.payment.Payment;
 import com.gestone.gestone_api.domain.payment.PaymentDTO;
 import com.gestone.gestone_api.domain.payment.PaymentStatus;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
 public record MarbleshopOrderDTO(
     String workAddress,
     UUID quotationId,
-    Integer discount,
+    BigDecimal discount,
     List<PaymentDTO> payments,
     String notes
 ) {
