+++ src/main/java/com/gestone/gestone_api/domain/order/MarbleshopOrderResponseDTO.java
@@ -18,7 +18,7 @@ public record MarbleshopOrderResponseDTO(
     CustomerResponseDTO customer,
     BigDecimal totalValue,
     BigDecimal totalArea,
-    Integer discount,
+    BigDecimal discount,
     BigDecimal finalValue,
     String marbleshopOrderStatus,
     List<MarbleshopItemResponseDTO> marbleshopItems,
